requirements.txt
alembic==1.7.5
click==8.0.3
Flask==2.0.2
Flask-Cors==3.0.10
Flask-JWT-Extended==4.3.1
Flask-Migrate==3.1.0
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
itsdangerous==2.0.1
Jinja2==3.0.3
Mako==1.1.6
MarkupSafe==2.0.1
marshmallow==3.14.1
PyJWT==2.3.0
python-dotenv==0.19.2
six==1.16.0
SQLAlchemy==1.4.31
Werkzeug==2.0.2
